+++ Changelog.md
@@ -3,7 +3,8 @@
 ## 0.18.0 (next)
 
 * `CursorRow::get_text` now returns `bool` with `false` indicating a `NULL` field.
-
+* `TextColumn` is now generic over its character type. The type alias `CharColumn` for
+  `TextColumn<u8>` can serve as a drop in replacement for the previous iteration of `TextColumn`.
 ## 0.17.0
 
 * Introduces `Cursor::next_row` to fetch data without binding buffers first.